// OneDrive/Bureau/--/pfe/app/src/navigation/types.ts
export type RootStackParamList = {
  Login: undefined;
  AppTabs: undefined;
  PatientDetail: { patientId: string };
};

export type AppTabParamList = {
  Dashboard: undefined;
  Patients: undefined;
  Alerts: undefined;
  Notifications: undefined;
  Simulation: undefined;
  Hospital: undefined;
  Users: undefined;
  Settings: undefined;
};